add AIAgentServiceBuilderBase which does not have .id() and .clock() setters
